+++ src/compas_rrc/__init__.py
@@ -56,6 +56,22 @@ The following example shows how to retrieve, update and send the robot joints an
     RobotJoints
     ExternalAxes
 
+# System instructions
+# -------------------
+
+# Besides the normal instructions, RRC can send and receive system-level instructions to the
+# controller.
+
+# .. autosummary::
+#     :toctree: generated/
+#     :nosignatures:
+
+#     GetControllerState
+#     GetExecutionState
+#     GetOperationMode
+#     GetSpeedRatio
+#     SetSpeedRatio
+
 Debugging instructions
 ----------------------
 
@@ -108,67 +124,108 @@ from compas_rrc.io import (
     SetDigital,
     SetGroup,
 )
-from compas_rrc.motion import Motion, MoveToFrame, MoveToJoints, MoveToRobtarget, Zone
+from compas_rrc.motion import (
+    Motion,
+    MoveToFrame,
+    MoveToJoints,
+    MoveToRobtarget,
+    Zone,
+)
 from compas_rrc.msg import PrintText
+from compas_rrc.system import (
+    GetControllerState,
+    GetExecutionState,
+    GetOperationMode,
+    GetSpeedRatio,
+    SetSpeedRatio,
+)
 from compas_rrc.utility import (
     Debug,
     GetFrame,
     GetJoints,
     GetRobtarget,
+    GetVariable,
     Noop,
+    ResetApp,
     SetAcceleration,
     SetMaxSpeed,
     SetTool,
+    SetVariable,
     SetWorkObject,
+    StartApp,
     Stop,
+    StopApp,
     WaitTime,
 )
-from compas_rrc.watch import ReadWatch, StartWatch, StopWatch
+from compas_rrc.watch import (
+    ReadWatch,
+    StartWatch,
+    StopWatch,
+)
 
 __all_plugins__ = ["compas_rrc.__install"]
 __all__ = [
+    # __version__
     "__url__",
     "__version__",
     "__author__",
     "__author_email__",
     "__license__",
     "__copyright__",
-    "CLIENT_PROTOCOL_VERSION",
-    "FeedbackLevel",
-    "ExecutionLevel",
-    "InstructionException",
-    "TimeoutException",
-    "FutureResult",
-    "ExternalAxes",
-    "RobotJoints",
-    "RosClient",
+    # client
     "AbbClient",
-    "SetDigital",
-    "SetAnalog",
-    "SetGroup",
+    "RosClient",
+    # common
+    "CLIENT_PROTOCOL_VERSION",
+    "ExecutionLevel",
+    "ExternalAxes",
+    "FeedbackLevel",
+    "FutureResult",
+    "InstructionException",
+    "RobotJoints",
+    "TimeoutException",
+    # custom
+    "CustomInstruction",
+    # io
     "PulseDigital",
     "ReadAnalog",
     "ReadDigital",
     "ReadGroup",
-    "Zone",
+    "SetAnalog",
+    "SetDigital",
+    "SetGroup",
+    # motion
     "Motion",
-    "MoveToJoints",
     "MoveToFrame",
+    "MoveToJoints",
     "MoveToRobtarget",
+    "Zone",
+    # msg
     "PrintText",
-    "CustomInstruction",
-    "Noop",
+    # system
+    "GetControllerState",
+    "GetExecutionState",
+    "GetOperationMode",
+    "GetSpeedRatio",
+    "SetSpeedRatio",
+    # utility
+    "Debug",
     "GetFrame",
     "GetJoints",
     "GetRobtarget",
-    "SetAcceleration",
-    "SetTool",
-    "SetMaxSpeed",
-    "Stop",
-    "WaitTime",
-    "SetWorkObject",
-    "Debug",
+    "GetVariable",
+    "Noop",
     "ReadWatch",
+    "ResetApp",
+    "SetAcceleration",
+    "SetMaxSpeed",
+    "SetTool",
+    "SetVariable",
+    "SetWorkObject",
+    "StartApp",
     "StartWatch",
+    "Stop",
+    "StopApp",
     "StopWatch",
+    "WaitTime",
 ]